e2e: organize performance tests into per-feature performance/ subfolders
Moves:
- notebooks-positron/notebook-performance.test.ts -> notebooks-positron/performance/render.test.ts
- notebooks-positron/notebook-kernel-startup.test.ts -> notebooks-positron/performance/kernel-startup.test.ts
- sessions/session-startup.test.ts -> sessions/performance/console-startup.test.ts

Scales as perf work grows: new perf tests drop into performance/ without
per-file naming decisions. Glob-friendly: tests/**/performance/ hits all
perf tests across features.

Describe titles aligned on the `<feature>: <aspect>` pattern:
- Positron Notebooks: Render
- Positron Notebooks: Kernel Startup
- Sessions: Console Startup
